requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
billiard==4.2.1
bleach==6.2.0
blinker==1.9.0
celery==5.4.0
celery-redbeat==2.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
DateTime==5.5
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
executing==2.2.0
fastjsonschema==2.21.1
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
flower==2.0.1
fqdn==1.5.1
frozenlist==1.5.0
gotrue==2.11.3
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.12.0
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
Js2Py==0.74
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kombu==5.4.2
langchain==0.3.17
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langsmith==0.3.6
lxml==5.3.0
m3u8==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
multidict==6.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
openai==1.61.1
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
postgrest==0.19.3
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyjsparser==2.7.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
realtime==2.3.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
Send2Trash==1.8.3
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
SQLAlchemy==2.0.37
stack-data==0.6.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tenacity==9.0.0
terminado==0.18.1
tiktoken==0.8.0
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
twikit==2.3.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
uri-template==1.3.0
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
webvtt-py==0.5.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
yarl==1.18.3
zope.interface==7.2
zstandard==0.23.0
